+++ pg_documentdb/include/commands/commands_common.h
@@ -124,6 +124,7 @@ bool FindShardKeyValueForDocumentId(MongoCollection *collection, const
 									const char *collationString);
 
 bool IsCommonSpecIgnoredField(const char *fieldName);
+void ValidateNamespaceStringForEmbeddedNull(const char *value, uint32_t length);
 void ValidateOrExtractDatabaseNameFromSpec(bson_iter_t *iter, Datum *databaseNameDatum);
 void ValidateOrExtractDatabaseNameTextFromSpec(bson_iter_t *iter,
 											   text **databaseNameText);